requirements.txt
annotated-types==0.7.0
anyio==4.11.0
click==8.3.0
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
fastapi==0.118.0
filelock==3.19.1
fonttools==4.60.1
fsspec==2025.9.0
grad-cam==1.5.5
h11==0.16.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.6
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
pydantic==2.11.10
pydantic_core==2.33.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-multipart==0.0.20
scikit-learn==1.7.2
scipy==1.16.2
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
tqdm==4.67.1
ttach==0.0.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0
